refactor(edit-engine): separate transaction protocol layers
