+++ packages/edit-engine/src/edit-schema.ts
@@ -1,0 +1,327 @@
+import {
+  AnnotationSchema,
+  DraftingObjectSchema,
+  InstanceNetlistDataSchema,
+  InstancePropertyValueSchema,
+  InstanceSchema,
+  JunctionRoleSchema,
+  LayoutConstraintSchema,
+  LayoutGroupSchema,
+  MirrorSchema,
+  NoConnectSchema,
+  PlacementSchema,
+  PointSchema,
+  RouteEndpointSchema,
+  RoutePresentationSchema,
+  RotationSchema,
+  SegmentModeSchema,
+  StableIdSchema,
+} from "@icm/model";
+import { z } from "zod";
+
+/**
+ * Public typed-edit protocol. This module owns payload shape only; execution,
+ * document lookup, and mutation invariants remain in `transaction.ts`.
+ */
+export const EditActorSchema = z.strictObject({
+  kind: z.enum(["human", "agent"]),
+  id: StableIdSchema,
+});
+
+export const NoopEditSchema = z.strictObject({
+  kind: z.literal("noop"),
+  reason: z.string().min(1).optional(),
+});
+export const ClearDocumentEditSchema = z.strictObject({
+  kind: z.literal("clear_document"),
+});
+export const AddInstanceEditSchema = z.strictObject({
+  kind: z.literal("add_instance"),
+  instance: InstanceSchema,
+});
+export const RemoveInstanceEditSchema = z.strictObject({
+  kind: z.literal("remove_instance"),
+  instanceId: StableIdSchema,
+});
+export const SetInstanceSymbolEditSchema = z.strictObject({
+  kind: z.literal("set_instance_symbol"),
+  instanceId: StableIdSchema,
+  symbolId: StableIdSchema,
+  symbolVariantId: StableIdSchema.nullable().optional(),
+  pinMap: z.record(z.string().min(1), z.string().min(1)).optional(),
+});
+export const PlaceInstanceEditSchema = z.strictObject({
+  kind: z.literal("place_instance"),
+  instanceId: StableIdSchema,
+  placement: PlacementSchema,
+});
+export const MoveInstanceEditSchema = z.strictObject({
+  kind: z.literal("move_instance"),
+  instanceId: StableIdSchema,
+  position: PointSchema,
+});
+export const RotateInstanceEditSchema = z.strictObject({
+  kind: z.literal("rotate_instance"),
+  instanceId: StableIdSchema,
+  rotation: RotationSchema,
+});
+export const MirrorInstanceEditSchema = z.strictObject({
+  kind: z.literal("mirror_instance"),
+  instanceId: StableIdSchema,
+  mirror: MirrorSchema,
+});
+export const PatchInstancePropertiesEditSchema = z
+  .strictObject({
+    kind: z.literal("patch_instance_properties"),
+    instanceId: StableIdSchema,
+    set: z.record(z.string().min(1), InstancePropertyValueSchema).optional(),
+    unset: z.array(z.string().min(1)).max(64).optional(),
+  })
+  .superRefine((edit, context) => {
+    for (const key of [...Object.keys(edit.set ?? {}), ...(edit.unset ?? [])]) {
+      if (!key.startsWith("spice.")) continue;
+      context.addIssue({
+        code: "custom",
+        path: [edit.set && key in edit.set ? "set" : "unset", key],
+        message:
+          "Legacy spice.* properties are migration-only; use typed netlist facts or import provenance",
+      });
+    }
+  });
+export const SetInstanceNetlistEditSchema = z.strictObject({
+  kind: z.literal("set_instance_netlist"),
+  instanceId: StableIdSchema,
+  netlist: InstanceNetlistDataSchema,
+});
+export const SetRoutePointsEditSchema = z.strictObject({
+  kind: z.literal("set_route_points"),
+  routeId: StableIdSchema,
+  netId: StableIdSchema,
+  from: RouteEndpointSchema,
+  to: RouteEndpointSchema,
+  waypoints: z.array(PointSchema),
+  segmentModes: z.array(SegmentModeSchema),
+  presentation: RoutePresentationSchema.optional(),
+});
+export const RouteOrthogonalEditSchema = z.strictObject({
+  kind: z.literal("route_orthogonal"),
+  routeId: StableIdSchema,
+  netId: StableIdSchema,
+  from: RouteEndpointSchema,
+  to: RouteEndpointSchema,
+  escapeLength: z.number().int().positive().max(1000).optional(),
+  presentation: RoutePresentationSchema.optional(),
+});
+export const AddJunctionEditSchema = z.strictObject({
+  kind: z.literal("add_junction"),
+  junctionId: StableIdSchema,
+  netId: StableIdSchema,
+  position: PointSchema,
+  role: JunctionRoleSchema.optional(),
+  createNet: z.boolean().optional(),
+  split: z
+    .strictObject({
+      routeId: StableIdSchema,
+      firstRouteId: StableIdSchema,
+      secondRouteId: StableIdSchema,
+      segmentIndex: z.number().int().nonnegative(),
+    })
+    .optional(),
+});
+export const AttachEndpointToRouteEditSchema = z.strictObject({
+  kind: z.literal("attach_endpoint_to_route"),
+  endpoint: RouteEndpointSchema,
+  routeId: StableIdSchema,
+  point: PointSchema,
+  segmentIndex: z.number().int().nonnegative(),
+  firstRouteId: StableIdSchema,
+  secondRouteId: StableIdSchema,
+});
+export const RemoveJunctionEditSchema = z.strictObject({
+  kind: z.literal("remove_junction"),
+  junctionId: StableIdSchema,
+});
+export const MoveJunctionEditSchema = z.strictObject({
+  kind: z.literal("move_junction"),
+  junctionId: StableIdSchema,
+  position: PointSchema,
+});
+export const MakeFlightlineEditSchema = z.strictObject({
+  kind: z.literal("make_flightline"),
+  routeId: StableIdSchema,
+});
+export const CutConnectionEditSchema = z.strictObject({
+  kind: z.literal("cut_connection"),
+  routeId: StableIdSchema,
+});
+export const ConnectEndpointsEditSchema = z.strictObject({
+  kind: z.literal("connect_endpoints"),
+  from: RouteEndpointSchema,
+  to: RouteEndpointSchema,
+  newNetId: StableIdSchema.optional(),
+  newNetName: z.string().min(1).optional(),
+  newNetScope: z.enum(["local", "global"]).optional(),
+});
+
+/** A VDD rail edit creates/reuses one explicit supply Net and its geometry. */
+export const AddPowerRailEditSchema = z.strictObject({
+  kind: z.literal("add_power_rail"),
+  netId: StableIdSchema,
+  routeId: StableIdSchema,
+  startJunctionId: StableIdSchema,
+  endJunctionId: StableIdSchema,
+  labelId: StableIdSchema,
+  domain: z.literal("vdd"),
+  start: PointSchema,
+  end: PointSchema,
+});
+export const SetNetPowerDomainEditSchema = z.strictObject({
+  kind: z.literal("set_net_power_domain"),
+  netId: StableIdSchema,
+  powerDomain: z.enum(["none", "vdd", "ground"]),
+});
+export const MergeNetsEditSchema = z.strictObject({
+  kind: z.literal("merge_nets"),
+  targetNetId: StableIdSchema,
+  sourceNetId: StableIdSchema,
+});
+export const SetNetNameEditSchema = z.strictObject({
+  kind: z.literal("set_net_name"),
+  netId: StableIdSchema,
+  name: z.string().trim().min(1).max(256),
+});
+export const NormalizePowerNetsEditSchema = z.strictObject({
+  kind: z.literal("normalize_power_nets"),
+});
+export const SetMosBulkDefaultsEditSchema = z.strictObject({
+  kind: z.literal("set_mos_bulk_defaults"),
+  nmosNetId: StableIdSchema.nullable().optional(),
+  pmosNetId: StableIdSchema.nullable().optional(),
+});
+export const ReconcileMosBulkEditSchema = z.strictObject({
+  kind: z.literal("reconcile_mos_bulk"),
+  instanceIds: z.array(StableIdSchema).optional(),
+});
+export const ClearMosBulkDefaultEditSchema = z.strictObject({
+  kind: z.literal("clear_mos_bulk_default"),
+  instanceId: StableIdSchema,
+});
+export const DisconnectEndpointEditSchema = z.strictObject({
+  kind: z.literal("disconnect_endpoint"),
+  endpoint: z.strictObject({
+    kind: z.literal("terminal"),
+    instanceId: StableIdSchema,
+    pinName: z.string().min(1),
+  }),
+});
+export const AddNoConnectEditSchema = z.strictObject({
+  kind: z.literal("add_no_connect"),
+  noConnect: NoConnectSchema,
+});
+export const RemoveNoConnectEditSchema = z.strictObject({
+  kind: z.literal("remove_no_connect"),
+  noConnectId: StableIdSchema,
+});
+export const SetPresentationStyleEditSchema = z.strictObject({
+  kind: z.literal("set_presentation_style"),
+  styleProfileId: StableIdSchema,
+});
+export const UpsertSchematicAnnotationEditSchema = z.strictObject({
+  kind: z.literal("upsert_schematic_annotation"),
+  annotation: AnnotationSchema,
+});
+export const RemoveSchematicAnnotationEditSchema = z.strictObject({
+  kind: z.literal("remove_schematic_annotation"),
+  annotationId: StableIdSchema,
+});
+export const UpsertDraftingObjectEditSchema = z.strictObject({
+  kind: z.literal("upsert_drafting_object"),
+  object: DraftingObjectSchema,
+});
+export const RemoveDraftingObjectEditSchema = z.strictObject({
+  kind: z.literal("remove_drafting_object"),
+  objectId: StableIdSchema,
+});
+export const SetLayoutGroupEditSchema = z.strictObject({
+  kind: z.literal("set_layout_group"),
+  group: LayoutGroupSchema,
+});
+export const RemoveLayoutGroupEditSchema = z.strictObject({
+  kind: z.literal("remove_layout_group"),
+  groupId: StableIdSchema,
+});
+export const SetLayoutConstraintEditSchema = z.strictObject({
+  kind: z.literal("set_layout_constraint"),
+  constraint: LayoutConstraintSchema,
+});
+export const RemoveLayoutConstraintEditSchema = z.strictObject({
+  kind: z.literal("remove_layout_constraint"),
+  constraintId: StableIdSchema,
+});
+export const AlignInstancesEditSchema = z.strictObject({
+  kind: z.literal("align_instances"),
+  instanceIds: z.array(StableIdSchema).min(2).max(64),
+  axis: z.enum(["x", "y"]),
+  coordinate: z.number().int().optional(),
+});
+export const UndoEditSchema = z.strictObject({ kind: z.literal("undo") });
+export const RedoEditSchema = z.strictObject({ kind: z.literal("redo") });
+
+export const SchematicEditSchema = z.discriminatedUnion("kind", [
+  NoopEditSchema,
+  ClearDocumentEditSchema,
+  AddInstanceEditSchema,
+  RemoveInstanceEditSchema,
+  SetInstanceSymbolEditSchema,
+  PlaceInstanceEditSchema,
+  MoveInstanceEditSchema,
+  RotateInstanceEditSchema,
+  MirrorInstanceEditSchema,
+  PatchInstancePropertiesEditSchema,
+  SetInstanceNetlistEditSchema,
+  SetRoutePointsEditSchema,
+  RouteOrthogonalEditSchema,
+  AddJunctionEditSchema,
+  AttachEndpointToRouteEditSchema,
+  RemoveJunctionEditSchema,
+  MoveJunctionEditSchema,
+  MakeFlightlineEditSchema,
+  CutConnectionEditSchema,
+  ConnectEndpointsEditSchema,
+  AddPowerRailEditSchema,
+  MergeNetsEditSchema,
+  SetNetNameEditSchema,
+  SetNetPowerDomainEditSchema,
+  NormalizePowerNetsEditSchema,
+  SetMosBulkDefaultsEditSchema,
+  ReconcileMosBulkEditSchema,
+  ClearMosBulkDefaultEditSchema,
+  DisconnectEndpointEditSchema,
+  AddNoConnectEditSchema,
+  RemoveNoConnectEditSchema,
+  SetPresentationStyleEditSchema,
+  UpsertSchematicAnnotationEditSchema,
+  RemoveSchematicAnnotationEditSchema,
+  UpsertDraftingObjectEditSchema,
+  RemoveDraftingObjectEditSchema,
+  SetLayoutGroupEditSchema,
+  RemoveLayoutGroupEditSchema,
+  SetLayoutConstraintEditSchema,
+  RemoveLayoutConstraintEditSchema,
+  AlignInstancesEditSchema,
+  UndoEditSchema,
+  RedoEditSchema,
+]);
+
+export const EditTransactionSchema = z.strictObject({
+  transactionId: StableIdSchema,
+  documentId: StableIdSchema,
+  expectedRevision: z.number().int().nonnegative(),
+  actor: EditActorSchema,
+  dryRun: z.boolean().optional(),
+  edits: z.array(SchematicEditSchema).min(1).max(256),
+});
+
+export type EditActor = z.infer<typeof EditActorSchema>;
+export type SchematicEdit = z.infer<typeof SchematicEditSchema>;
+export type EditTransaction = z.infer<typeof EditTransactionSchema>;
